Changed hashrate type from long to double Added support for miners that don't return total hashrate

+++ Msv.AutoMiner/Msv.AutoMiner.Common/Models/CoinInfoService/AlgorithmData.cs
@@ -6,6 +6,6 @@ namespace Msv.AutoMiner.Common.Models.CoinInfoService
     {
         public Guid AlgorithmId { get; set; }
 
-        public long NetHashRate { get; set; }
+        public double NetHashRate { get; set; }
     }
 }
